requirements.txt
jsonpickle
trio==0.14.0
typing_extensions
pyipv8 @ git+https://github.com/Tribler/py-ipv8@2.14
pyyaml
click
networkx
matplotlib
pandas
pybloom_live